# docs/external-event-api.openapi.yaml
openapi: 3.1.0
info:
  title: Dance Billboard External Event API
  version: 1.3.0
  description: 외부 파트너가 소유한 Dance Billboard 일정을 등록·수정·삭제하는 서버 간 API
servers:
  - url: https://swingenjoy.com
paths:
...
  /api/external/v1/regular-socials:
    post:
      summary: 정규 소셜 반복 규칙 등록 또는 동일 ID 갱신
      operationId: createRegularSocialRule
      security:
        - partnerApiKey: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegularSocialRuleRequest'
      responses:
        '201':
          description: 규칙 등록 성공
        '200':
          description: 동일 규칙 갱신 또는 테스트 검증 성공
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '429':
          $ref: '#/components/responses/RateLimited'
  /api/external/v1/regular-socials/{external_id}:
    parameters:
      - name: external_id
        in: path
        required: true
        schema: { type: string, maxLength: 160 }
    put:
      summary: 정규 소셜 반복 규칙 수정
      operationId: updateRegularSocialRule
      security:
        - partnerApiKey: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegularSocialRuleRequest'
      responses:
        '200': { description: 수정 성공 }
        '404':
          $ref: '#/components/responses/NotFound'
    delete:
      summary: 정규 소셜 반복 규칙과 연결 예외 삭제
      operationId: deleteRegularSocialRule
      security:
        - partnerApiKey: []
      responses:
        '200': { description: 삭제 성공 }
        '404':
          $ref: '#/components/responses/NotFound'
...
  /api/external/v1/regular-socials/{external_id}/exceptions/{exception_id}:
    parameters:
      - name: external_id
        in: path
        required: true
        description: 반복 규칙 external_id
        schema: { type: string, maxLength: 160 }
      - name: exception_id
        in: path
        required: true
        schema: { type: string, maxLength: 160 }
    put:
      summary: 날짜별 예외 수정
      operationId: updateRegularSocialException
      security:
        - partnerApiKey: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegularSocialExceptionRequest'
      responses:
        '200': { description: 수정 성공 }
        '404':
          $ref: '#/components/responses/NotFound'
    delete:
      summary: 날짜별 예외 삭제
      operationId: deleteRegularSocialException
      security:
        - partnerApiKey: []
      responses:
        '200': { description: 삭제 성공 }
        '404':
          $ref: '#/components/responses/NotFound'
...
components:
  securitySchemes:
    partnerApiKey:
      type: http
      scheme: bearer
      bearerFormat: SwingEnjoyPartnerKey
  schemas:
    RegularSocialRuleRequest:
      type: object
      additionalProperties: false
      required: [external_id, title, weekday, location, source_url]
      properties:
        external_id: { type: string, maxLength: 160 }
        title: { type: string, maxLength: 255 }
        weekday:
          type: integer
          minimum: 0
          maximum: 6
          description: 일요일 0, 토요일 6
        time: { type: string, maxLength: 255 }
        location: { type: string, maxLength: 255 }
        venue_name: { type: string, maxLength: 255 }
        source_url: { type: string, format: uri, pattern: '^https://' }
        source_id: { type: string, maxLength: 160 }
        valid_from: { type: string, format: date }
        valid_until: { type: string, format: date }
        active: { type: boolean, default: true }
    RegularSocialExceptionRequest:
      type: object
      additionalProperties: false
      required: [external_id, date, type]
      properties:
        external_id: { type: string, maxLength: 160 }
        date: { type: string, format: date }
        type:
          type: string
          enum: [closure, override]
        title: { type: string, maxLength: 255 }
        time: { type: string, maxLength: 255 }
        location: { type: string, maxLength: 255 }
        venue_name: { type: string, maxLength: 255 }
        dj_name: { type: string, maxLength: 255 }
        source_url: { type: string, format: uri, pattern: '^https://' }
        description: { type: string, maxLength: 20000 }
...
    ErrorResponse:
      type: object
      required: [error, message]
      properties:
        error:
          type: string
        code:
          type: string
        message:
          type: string
  responses:
    BadRequest:
      description: 요청값 오류
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Unauthorized:
      description: API Key 오류
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    NotFound:
      description: 해당 파트너 소유 일정 없음
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
    RateLimited:
      description: 호출량 제한 초과
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'